Fix control lookup and eliminate some duplication in validation code
