+++ flucoma/doc/validate/object.py
@@ -6,13 +6,17 @@
 # under the European Union’s Horizon 2020 research and innovation programme
 # (grant agreement No 725899).
 
+
+import logging 
+from functools import partial, reduce
+
+from schema import Schema, Or, Optional, Use, SchemaError
+
 from .common import PermissveSchema, not_yet_documented
 from .controls import validate_controls
 from .messages import validate_messages
-from schema import Schema, Or, Optional, Use, SchemaError
-from functools import partial, reduce
 from ..defaults import DefaultControlDocs, DefaultMessageDocs
-import logging 
+
 from ..logger import add_context
 
 """
